Fixed NMR implementation and examples (#235)
* Fixed NMR implementation and examples


* Fixed NMR example styles


* Fixed NMR implementation style


* Fixed example bugs


* Convert NMR perspective rendering params to cuda automatically
