+++ .gitignore
@@ -196,6 +196,7 @@ public/
 
 # Results directory for examples
 examples/renderers/DIB-R/results/
+examples/renderers/NMR/results/
 
 # example outputs
 cache/
