+++ examples/renderers/NMR/README.md
@@ -1,63 +1,57 @@
-# Neural 3D Mesh Renderer (CVPR 2018)
+# Neural Mesh Renderer (NMR) Examples
 
-This repo contains a PyTorch implementation of the paper [Neural 3D Mesh Renderer](http://hiroharu-kato.com/projects_en/neural_renderer.html) by Hiroharu Kato, Yoshitaka Ushiku, and Tatsuya Harada.
-It is a port of the [original Chainer implementation](https://github.com/hiroharu-kato/neural_renderer) released by the authors.
-Currently the API is the same as in the original implementation with some smalls additions (e.g. render using a general 3x4 camera matrix, lens distortion coefficients etc.). However it is possible that it will change in the future.
+![](images/example3_optimization.gif)
 
-The library is fully functional and it passes all the test cases supplied by the authors of the original library.
-Detailed documentation will be added in the near future.
-## Requirements
-Python 2.7+ and PyTorch 0.4.0.
+These examples make use of the Neural Mesh Renderer in kaolin to render meshes and perform optimizations.
 
-The code has been tested only with PyTorch 0.4.0, there are no guarantees that it is compatible with older versions.
-Currently the library has both Python 3 and Python 2 support.
+Each of these examples will, by default, operate on the `rocket.obj` (and some of the PNG images) in this directory, and output animated GIF images to a `results` folder (which will be created if not exist).
 
-**Note**: In some newer PyTorch versions you might see some compilation errors involving AT_ASSERT. In these cases you can use the version of the code that is in the branch *at_assert_fix*. These changes will be merged into master in the near future.
-## Installation
-You can install the package by running
-```
-pip install neural_renderer_pytorch
-```
-Since running install.py requires PyTorch, make sure to install PyTorch before running the above command.
-## Running examples
-```
-python ./examples/example1.py
-python ./examples/example2.py
-python ./examples/example3.py
-python ./examples/example4.py
-```
+To run each example, simply execute `python <example-script>`, where `<example-script>` is one of the four examples in this directory.
 
+To see the command line arguments of each example, run `python <example-script> --help`.
 
-## Example 1: Drawing an object from multiple viewpoints
+## Example 1
 
-![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example1.gif)
+Renders the mesh from multiple viewpoints.
 
-## Example 2: Optimizing vertices
+Example output:
 
-Transforming the silhouette of a teapot into a rectangle. The loss function is the difference between the rendered image and the reference image.
+![](images/example1.gif)
 
-Reference image, optimization, and the result.
+## Example 2
 
-![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example2_ref.png) ![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example2_optimization.gif) ![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example2_result.gif)
+Transforming the silhouette of the mesh to match the target image (a square by default), by optimizing the vertices.
 
-## Example 3: Optimizing textures
+You can try different target images in this folder as well, such as `python <example-script> --image ./example2_moon.png` or `./example2_circle.png`.
 
-Matching the color of a teapot with a reference image.
+Example target image, final mesh, and optimization process:
 
-Reference image, result.
+![](example2_ref.png)
+![](images/example2_mesh.gif)
+![](images/example2_optimization.gif)
 
-![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example3_ref.png) ![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example3_result.gif)
+## Example 3
 
-## Example 4: Finding camera parameters
+Transforming the color of the mesh to match a target image, by optimizing the texture.
 
-The derivative of images with respect to camera pose can be computed through this renderer. In this example the position of the camera is optimized by gradient descent.
+Example target image, final mesh, and optimization process:
 
-From left to right: reference image, initial state, and optimization process.
+![](example3_ref.png)
+![](images/example3_mesh.gif)
+![](images/example3_optimization.gif)
 
-![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example4_ref.png) ![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example4_init.png) ![](https://raw.githubusercontent.com/hiroharu-kato/neural_renderer/master/examples/data/example4_result.gif)
+## Example 4
 
+Transform the camera position to match a target image, by optimizing it directly.
+
+Example target image and final mesh:
+
+![](example4_ref.png)
+![](images/example4.gif)
 
 ## Citation
+
+The Neural Mesh Renderer and the examples are based on its original implementation by Kato et al.
 
 ```
 @InProceedings{kato2018renderer
