--- examples/renderers/NMR/examples/example3.py
@@ -1,116 +1,0 @@
-# MIT License
-
-# Copyright (c) 2017 Hiroharu Kato
-# Copyright (c) 2018 Nikos Kolotouros
-# A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""
-Example 3. Optimizing textures.
-"""
-from __future__ import division
-import os
-import argparse
-import glob
-
-import torch
-import torch.nn as nn
-import numpy as np
-from skimage.io import imread, imsave
-import tqdm
-import imageio
-
-import neural_renderer as nr
-
-current_dir = os.path.dirname(os.path.realpath(__file__))
-data_dir = os.path.join(current_dir, 'data')
-
-class Model(nn.Module):
-    def __init__(self, filename_obj, filename_ref):
-        super(Model, self).__init__()
-        vertices, faces = nr.load_obj(filename_obj)
-        self.register_buffer('vertices', vertices[None, :, :])
-        self.register_buffer('faces', faces[None, :, :])
-
-        # create textures
-        texture_size = 4
-        textures = torch.zeros(1, self.faces.shape[1], texture_size, texture_size, texture_size, 3, dtype=torch.float32)
-        self.textures = nn.Parameter(textures)
-
-        # load reference image
-        image_ref = torch.from_numpy(imread(filename_ref).astype('float32') / 255.).permute(2,0,1)[None, ::]
-        self.register_buffer('image_ref', image_ref)
-
-        # setup renderer
-        renderer = nr.Renderer(camera_mode='look_at')
-        renderer.perspective = False
-        renderer.light_intensity_directional = 0.0
-        renderer.light_intensity_ambient = 1.0
-        self.renderer = renderer
-
-
-    def forward(self):
-        self.renderer.eye = nr.get_points_from_angles(2.732, 0, np.random.uniform(0, 360))
-        image, _, _ = self.renderer(self.vertices, self.faces, torch.tanh(self.textures))
-        loss = torch.sum((image - self.image_ref) ** 2)
-        return loss
-
-
-def make_gif(filename):
-    with imageio.get_writer(filename, mode='I') as writer:
-        for filename in sorted(glob.glob('/tmp/_tmp_*.png')):
-            writer.append_data(imageio.imread(filename))
-            os.remove(filename)
-    writer.close()
-
-
-def main():
-    parser = argparse.ArgumentParser()
-    parser.add_argument('-io', '--filename_obj', type=str, default=os.path.join(data_dir, 'teapot.obj'))
-    parser.add_argument('-ir', '--filename_ref', type=str, default=os.path.join(data_dir, 'example3_ref.png'))
-    parser.add_argument('-or', '--filename_output', type=str, default=os.path.join(data_dir, 'example3_result.gif'))
-    parser.add_argument('-g', '--gpu', type=int, default=0)
-    args = parser.parse_args()
-
-    model = Model(args.filename_obj, args.filename_ref)
-    model.cuda()
-
-    optimizer = torch.optim.Adam(model.parameters(), lr=0.1, betas=(0.5,0.999))
-    loop = tqdm.tqdm(range(300))
-    for _ in loop:
-        loop.set_description('Optimizing')
-        optimizer.zero_grad()
-        loss = model()
-        loss.backward()
-        optimizer.step()
-
-    # draw object
-    loop = tqdm.tqdm(range(0, 360, 4))
-    for num, azimuth in enumerate(loop):
-        loop.set_description('Drawing')
-        model.renderer.eye = nr.get_points_from_angles(2.732, 0, azimuth)
-        images, _, _ = model.renderer(model.vertices, model.faces, torch.tanh(model.textures))
-        image = images.detach().cpu().numpy()[0].transpose((1, 2, 0))
-        imsave('/tmp/_tmp_%04d.png' % num, image)
-    make_gif(args.filename_output)
-
-
-if __name__ == '__main__':
-    main()
