--- examples/renderers/NMR/neural_renderer/__init__.py
@@ -1,39 +1,0 @@
-# MIT License
-
-# Copyright (c) 2017 Hiroharu Kato
-# Copyright (c) 2018 Nikos Kolotouros
-# A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from .get_points_from_angles import get_points_from_angles
-from .lighting import lighting
-from .load_obj import load_obj
-from .look import look
-from .look_at import look_at
-from .mesh import Mesh
-from .perspective import perspective
-from .projection import projection
-from .rasterize import (rasterize_rgbad, rasterize, rasterize_silhouettes, rasterize_depth, Rasterize)
-from .renderer import Renderer
-from .save_obj import save_obj
-from .vertices_to_faces import vertices_to_faces
-
-__version__ = '1.1.3'
-name = 'neural_renderer_pytorch'
