--- examples/renderers/NMR/neural_renderer/cuda/load_textures_cuda.cpp
@@ -1,63 +1,0 @@
-// MIT License
-
-// Copyright (c) 2017 Hiroharu Kato
-// Copyright (c) 2018 Nikos Kolotouros
-// A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-
-// The above copyright notice and this permission notice shall be included in all
-// copies or substantial portions of the Software.
-
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-// SOFTWARE.
-
-#include <torch/torch.h>
-
-// CUDA forward declarations
-
-at::Tensor load_textures_cuda(
-        at::Tensor image,
-        at::Tensor faces,
-        at::Tensor textures,
-        at::Tensor is_update,
-        int texture_wrapping,
-        int use_bilinear);
-
-// C++ interface
-
-#define CHECK_CUDA(x) AT_CHECK(x.type().is_cuda(), #x " must be a CUDA tensor")
-#define CHECK_CONTIGUOUS(x) AT_CHECK(x.is_contiguous(), #x " must be contiguous")
-#define CHECK_INPUT(x) CHECK_CUDA(x); CHECK_CONTIGUOUS(x)
-
-
-at::Tensor load_textures(
-        at::Tensor image,
-        at::Tensor faces,
-        at::Tensor textures,
-        at::Tensor is_update,
-        int texture_wrapping,
-        int use_bilinear) {
-
-    CHECK_INPUT(image);
-    CHECK_INPUT(faces);
-    CHECK_INPUT(is_update);
-    CHECK_INPUT(textures);
-
-    return load_textures_cuda(image, faces, textures, is_update, texture_wrapping, use_bilinear);
-                                      
-}
-
-PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
-    m.def("load_textures", &load_textures, "LOAD_TEXTURES (CUDA)");
-}
