--- examples/renderers/NMR/neural_renderer/get_points_from_angles.py
@@ -1,48 +1,0 @@
-# MIT License
-
-# Copyright (c) 2017 Hiroharu Kato
-# Copyright (c) 2018 Nikos Kolotouros
-# A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import division
-import math
-
-import torch
-
-def get_points_from_angles(distance, elevation, azimuth, degrees=True):
-    if isinstance(distance, float) or isinstance(distance, int):
-        if degrees:
-            elevation = math.radians(elevation)
-            azimuth = math.radians(azimuth)
-        return (
-            distance * math.cos(elevation) * math.sin(azimuth),
-            distance * math.sin(elevation),
-            -distance * math.cos(elevation) * math.cos(azimuth))
-    else:
-        if degrees:
-            elevation = math.pi/180. * elevation
-            azimuth = math.pi/180. * azimuth
-    #
-        return torch.stack([
-            distance * torch.cos(elevation) * torch.sin(azimuth),
-            distance * torch.sin(elevation),
-            -distance * torch.cos(elevation) * torch.cos(azimuth)
-            ]).transpose(1,0)
