--- examples/renderers/NMR/neural_renderer/look.py
@@ -1,77 +1,0 @@
-# MIT License
-
-# Copyright (c) 2017 Hiroharu Kato
-# Copyright (c) 2018 Nikos Kolotouros
-# A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-import numpy as np
-import torch
-import torch.nn.functional as F
-
-
-def look(vertices, eye, direction=[0, 1, 0], up=None):
-    """
-    "Look" transformation of vertices.
-    """
-    if (vertices.ndimension() != 3):
-        raise ValueError('vertices Tensor should have 3 dimensions')
-
-    device = vertices.device
-
-    if isinstance(direction, list) or isinstance(direction, tuple):
-        direction = torch.tensor(direction, dtype=torch.float32, device=device)
-    elif isinstance(direction, np.ndarray):
-        direction = torch.from_numpy(direction).to(device)
-    elif torch.is_tensor(direction):
-        direction = direction.to(device)
-
-    if isinstance(eye, list) or isinstance(eye, tuple):
-        eye = torch.tensor(eye, dtype=torch.float32, device=device)
-    elif isinstance(eye, np.ndarray):
-        eye = torch.from_numpy(eye).to(device)
-    elif torch.is_tensor(eye):
-        eye = eye.to(device)
-
-    if up is None:
-        up = torch.cuda.FloatTensor([0, 1, 0])
-    if eye.ndimension() == 1:
-        eye = eye[None, :]
-    if direction.ndimension() == 1:
-        direction = direction[None, :]
-    if up.ndimension() == 1:
-        up = up[None, :]
-
-    # create new axes
-    z_axis = F.normalize(direction, eps=1e-5)
-    x_axis = F.normalize(torch.cross(up, z_axis), eps=1e-5)
-    y_axis = F.normalize(torch.cross(z_axis, x_axis), eps=1e-5)
-
-    # create rotation matrix: [bs, 3, 3]
-    r = torch.cat((x_axis[:, None, :], y_axis[:, None, :], z_axis[:, None, :]), dim=1)
-
-    # apply
-    # [bs, nv, 3] -> [bs, nv, 3] -> [bs, nv, 3]
-    if vertices.shape != eye.shape:
-        eye = eye[:, None, :]
-    vertices = vertices - eye
-    vertices = torch.matmul(vertices, r.transpose(1,2))
-
-    return vertices
