--- examples/renderers/NMR/neural_renderer/perspective.py
@@ -1,45 +1,0 @@
-# MIT License
-
-# Copyright (c) 2017 Hiroharu Kato
-# Copyright (c) 2018 Nikos Kolotouros
-# A PyTorch implementation of Neural 3D Mesh Renderer (https://github.com/hiroharu-kato/neural_renderer)
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-from __future__ import division
-import math
-
-import torch
-
-def perspective(vertices, angle=30.):
-    '''
-    Compute perspective distortion from a given angle
-    '''
-    if (vertices.ndimension() != 3):
-        raise ValueError('vertices Tensor should have 3 dimensions')
-    device = vertices.device
-    angle = torch.tensor(angle / 180 * math.pi, dtype=torch.float32, device=device)
-    angle = angle[None]
-    width = torch.tan(angle)
-    width = width[:, None] 
-    z = vertices[:, :, 2]
-    x = vertices[:, :, 0] / z / width
-    y = vertices[:, :, 1] / z / width
-    vertices = torch.stack((x,y,z), dim=2)
-    return vertices
