+++ kaolin/graphics/nmr/cuda/rasterize_cuda_kernel.cu
@@ -29,9 +29,18 @@
 #include <cuda_runtime.h>
 
 // for the older gpus atomicAdd with double arguments does not exist
-#if !defined(__CUDA_ARCH__) || __CUDA_ARCH__ >= 600
-#else
-__device__ double atomicAdd(double* a, double b) { return b; }
+#if  __CUDA_ARCH__ < 600 and defined(__CUDA_ARCH__)
+static __inline__ __device__ double atomicAdd(double* address, double val) {
+    unsigned long long int* address_as_ull = (unsigned long long int*)address;
+    unsigned long long int old = *address_as_ull, assumed;
+    do {
+        assumed = old;
+        old = atomicCAS(address_as_ull, assumed,
+                __double_as_longlong(val + __longlong_as_double(assumed)));
+    // Note: uses integer comparison to avoid hang in case of NaN (since NaN != NaN) } while (assumed != old);
+    } while (assumed != old);
+    return __longlong_as_double(old);
+}
 #endif
 
 namespace{
